fixed crash if total number of subkeys changes during RegEnumKeys call
It's theoretically possible for the registry to change while we
enumerate through the subkeys of a HKEY - in which case we would receive
ERROR_NO_MORE_ITEMS instead of ERROR_SUCCESS which simply signifies that
there's no data left to query.

Previously we would throw an exception if RegEnumKeyExW's result was
different than ERROR_SUCCESS, causing us to lose the result array which
would definitely hold valid entries when ERROR_NO_MORE_ITEMS is reported.

fixes nexus-mods/vortex#7675